+++ cmd/werf/common/common.go
@@ -756,7 +756,7 @@ func SetupFollow(cmdData *CmdData, cmd *cobra.Command) {
 	cmdData.Follow = new(bool)
 	cmd.Flags().BoolVarP(cmdData.Follow, "follow", "", GetBoolEnvironmentDefaultFalse("WERF_FOLLOW"), `Enable follow mode (default $WERF_FOLLOW).
 The mode allows restarting the command on a new commit.
-In development mode (--dev), it additionally tracks changes in the index state of the git repository`)
+In development mode (--dev), werf restarts the command on any changes (including untracked files) in the git repository worktree`)
 }
 
 func allStagesNames() []string {