+++ docs/documentation/pages_en/whats_new_in_v1_2/changelog.md
@@ -26,7 +26,7 @@ By default, each of these commands reads all needed files from the current commi
 
 Command with the `--follow` flag will work in a loop, where either:
  - a new commit to the project git work tree will cause command rerun — by default;
- - changes to the git index of the project git work tree will cause command rerun —  when `--follow` flag is combined with the `--dev` flag.
+ - any changes in the git repository worktree including untracked files will cause command rerun —  when `--follow` flag is combined with the `--dev` flag.
 
 Internally werf will commit changes to dev branch `werf-dev-<commit>`. Dev-cache is linked only to those temporal commits and will not interfere with the main cache (when no `--dev` has been specified).